refactor(daemon): split wire prost module
Move daemon wire prost codecs into focused submodules while preserving the existing daemon-facing API.

Refs #332
